Major refactoring in classes Versatile, Parametre and Swig wrappings. Add new boolean parameter "EXCLUSIVE". Implement parameter "EXCLUSIVE" for ePBS, eLSF, eSlurm and eLL. Some other fixes and improvements.

+++ src/CCC/Batch_BatchManager_eCCC.cxx
@@ -243,7 +243,7 @@ namespace Batch {
     if (params.find(NBPROC) != params.end()) 
       nbproc = params[NBPROC];
     if (params.find(MAXWALLTIME) != params.end()) 
-      edt = params[MAXWALLTIME] * 60;
+      edt = (long)params[MAXWALLTIME] * 60;
     if (params.find(MAXRAMSIZE) != params.end()) 
       mem = params[MAXRAMSIZE];
     if (params.find(QUEUE) != params.end())
